% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RABIT.R
\docType{package}
\name{RABIT}
\alias{RABIT}
\alias{RABIT-package}
\title{\code{RABIT}: Regression Analysis with Background InTegration}
\description{
The \code{RABIT} package is a powerful feature selection algortihm
}

