% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hellow}
\alias{rcpp_hellow}
\title{rcpp_hellow()}
\usage{
rcpp_hellow(q)
}
\arguments{
\item{q}{if we parameterize the function}
}
\value{
Returns a list object
}
\description{
Prints basic
}
\details{
\code{rcpp_hello} does basic calling
}

