% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runRABIT}
\alias{runRABIT}
\title{runRABIT}
\usage{
runRABIT(x, y, b = matrix(2), c = matrix(2), f = 0.05, t = TRUE,
  s = FALSE, r = TRUE)
}
\arguments{
\item{x}{Input variable matrix. Each column represents a variable to be selected.
REQUIRED.}

\item{y}{Response matrix. Each column represents a response vector and Rabit will
run variable selection on each response separately. REQUIRED.}

\item{b}{Background factors. Each column represents a confounding factor to be controlled.
Rabit will exclude these effects from the linear model. Default = NULL}

\item{c}{Background factors for each column of Y. Different from b which inputs background
factors shared across all columns of Y, this input is specific for each column of Y. Rabit
will search for shared column names between Y and input background factors here and control
these factors. default = NULL}

\item{f}{FDR threshold. Range (0,1]. Rabit estimates the statistical significance of each
variable in X and only keeps significant ones in later stepwise forward regression. Set
to 1 to skip this step. default = 0.05}

\item{t}{Transform Y to normal distribution ( = TRUE) or not ( = FALSE). Transforming response
Y to normal distribution increases the statistical power of linear regression t-test. However,
if Y is very different from the normal distribution (e.g. binary outcome), then this
methodology should not be used and logistic regression should be used in this example.
default = TRUE}

\item{s}{Select one best variable in X for each category ( = TRUE) or not ( = FALSE).
 When several  variables in X come from the same category (e.g. Transcription factors
may have multiple ChIP-Seq profiles available), multiple names appended together
with a . (e.g. CatA.V1, CatA.V2 ...) Rabit will chose the best one based on the
maximum t-value and these variableswill be returned. default = FALSE}

\item{r}{Run forward stepwise selection ( = TRUE) or not ( = FALSE). If set to FALSE,
Rabit will only calculate the statistical significance of each variable without
forward stepwise selection. default = TRUE.}
}
\value{
Who knows but hopefully it's okay.
}
\description{
Performs regression analysis
}
\details{
\code{runRABIT} as a function call is a one-stop shop. Carefully consider each parameter before
proceeding.
}

